+++ src/Altinn.App.Api/Controllers/DataController.cs
@@ -53,7 +53,6 @@ namespace Altinn.App.Api.Controllers
         /// <param name="logger">logger</param>
         /// <param name="instanceClient">instance service to store instances</param>
         /// <param name="dataClient">A service with access to data storage.</param>
-        /// <param name="altinnApp">The app logic for current service</param>
         /// <param name="appResourcesService">The apps resource service</param>
         /// <param name="prefillService">A service with prefill related logic.</param>
         public DataController(
