+++ src/Altinn.App.Api/Controllers/OptionsController.cs
@@ -4,7 +4,6 @@ using System.Threading.Tasks;
 using Altinn.App.Common.Models;
 using Altinn.App.PlatformServices.Models;
 using Altinn.App.PlatformServices.Options;
-using Altinn.App.Services.Interface;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -23,7 +22,6 @@ namespace Altinn.App.Api.Controllers
         /// <summary>
         /// Initializes a new instance of the <see cref="OptionsController"/> class.
         /// </summary>
-        /// <param name="altinnApp">The current App Core used to interface with custom logic</param>
         /// <param name="appOptionsService">Service for handling app options</param>
         public OptionsController(IAppOptionsService appOptionsService)
         {
