+++ src/Altinn.App.Core/EFormidling/Implementation/DefaultEFormidlingReceivers.cs
@@ -2,19 +2,26 @@ using Altinn.App.Core.EFormidling.Interface;
 using Altinn.App.Services.Interface;
 using Altinn.Common.EFormidlingClient.Models.SBD;
 using Altinn.Platform.Storage.Interface.Models;
-using Microsoft.Rest.Azure;
 
 namespace Altinn.App.Core.EFormidling.Implementation;
 
+/// <summary>
+/// Default implementation of <see cref="Altinn.App.Core.EFormidling.Interface.IEFormidlingReceivers"/>
+/// </summary>
 public class DefaultEFormidlingReceivers: IEFormidlingReceivers
 {
     private readonly Application _appMetadata;
     
+    /// <summary>
+    /// Initializes a new instance of the <see cref="DefaultEFormidlingReceivers"/> class.
+    /// </summary>
+    /// <param name="resources"></param>
     public DefaultEFormidlingReceivers(IAppResources resources)
     {
         _appMetadata = resources.GetApplication();
     }
     
+    /// <inheritdoc />
     public async Task<List<Receiver>> GetEFormidlingReceivers(Instance instance)
     {
         await Task.CompletedTask;
